app/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
mangum==0.17.0

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# File handling
aiofiles==23.2.1

# Date/time utilities
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
